requirements.txt
amqp==5.0.6
asgiref==3.3.4
astroid==2.5.6
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
coveralls==3.0.1
Django==3.2
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-guardian==2.3.0
django-redis==4.12.1
djangorestframework==3.12.4
djangorestframework-bulk==0.2.1
djangorestframework-guardian==0.3.0
djangorestframework-jwt==1.11.0
docopt==0.6.2
flake8==3.9.2
idna==2.10
importlib-metadata==4.0.1
isort==5.8.0
itypes==1.2.0
Jinja2==3.0.0
kombu==5.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.0
mccabe==0.6.1
mysqlclient==2.0.3
prompt-toolkit==3.0.18
pycodestyle==2.7.0
pycrypto==2.6.1
pyflakes==2.3.1
PyJWT==1.7.1
pylint==2.8.2
pylint-django==2.4.4
pylint-plugin-utils==0.6
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.1
